move code from Diagnostics and Guard to cpp
